add a simple test to see if it does not throw an error
