+++ __tests__/index.tests.ts
@@ -1,0 +1,12 @@
+// import * as core from '@actions/core'
+import extractJiraKeysFromCommit from '../index'
+
+beforeEach( () => {
+    jest.resetModules();
+})
+
+describe('debug action debug messages', () => {
+  it('extract Jira Keys From Commit', async () => {
+    await expect(extractJiraKeysFromCommit()).resolves.not.toThrow()
+  })
+})
