--- debug-action/__tests__/debug.tests.ts
@@ -1,10 +1,0 @@
-import * as core from '@actions/core'
-import run from '../debug'
-
-describe('debug action debug messages', () => {
-  it('outputs a debug message', async () => {
-    const debugMock = jest.spyOn(core, 'debug')
-    await run()
-    expect(debugMock).toHaveBeenCalledWith('👋 Hello! You are an amazing person! 🙌')
-  })
-})
