--- debug-action/debug.ts
@@ -1,9 +1,0 @@
-import * as core from '@actions/core'
-
-const run = async (): Promise<void> => {
-    core.debug('👋 Hello! You are an amazing person! 🙌')
-}
-
-run()
-
-export default run
